refactor: 사용하지 않는 dto 삭제
