--- backend/src/main/java/org/example/backend/users/domain/dto/member/SjLoginReq.java
@@ -1,24 +1,0 @@
-package org.example.backend.users.domain.dto.member;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class SjLoginReq {
-    private String userId;
-    private String password;
-
-    public SjLoginReq(String userId, String password) {
-        this.userId = userId;
-        this.password = password;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
