+++ backend/src/main/java/org/example/backend/users/service/AdminService.java
@@ -6,7 +6,7 @@ import static org.example.backend.users.exception.admin.AdminExceptionType.INVAL
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.backend.users.domain.dto.admin.AccessTokenReq;
-import org.example.backend.users.domain.dto.admin.SignInReqDto;
+import org.example.backend.users.domain.dto.SignInReqDto;
 import org.example.backend.users.domain.entity.Role;
 import org.example.backend.users.domain.entity.Users;
 import org.example.backend.users.exception.admin.AdminException;
